DZIGA

Documentation for DZIGA.

DZIGA.CityGraph

DZIGA.get_upper_bound

DZIGA.greedy_planner

DZIGA.CityGraph — Type

CityGraph

story a city as a graph adjacency matrix for easy traversal. Given a 'City', produce a matrix dictionary with {endpointA : {endpointB : (duration, distance)} as key-value pairs. Also create an edges vector of tuples of (endpointA, endpointB) for easy traversal.

Fields

matrix::Dict{Int,Dict{Int,Tuple{Int,Int}}}: a dictionary of dictionaries where the first key is the starting junction and the second key is the ending junction. The value is a tuple of the duration and distance of the street.

edges::Vector{Tuple{Int,Int}}: a vector of tuples of the starting and ending junctions of the street.

DZIGA.get_upper_bound — Method

get_upper_bound(city::City) ::Int

Return the greatest possible distance you can traverse for the given City by summing the longest streets first.

DZIGA.greedy_planner — Method

greedy_planner(city::City) ::Int

Greedily the greatest possible distance you can traverse for the given City. Best algorithm we have.